+++ packages/abjad-web/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  ...require('../../.eslintrc.js'),
-  extends: [
-    'plugin:@next/next/recommended',
-  ],
-}