+++ packages/abjad-web/src/app/layout.tsx
@@ -1,6 +1,5 @@
 import Head from 'next/head'
 import { Analytics } from '@vercel/analytics/react'
-import { SpeedInsights } from '@vercel/speed-insights/next'
 import { Inter } from 'next/font/google'
 import type { Metadata } from 'next'
 import React from 'react'
@@ -26,7 +25,6 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
 			<body className={inter.className}>
 				{children}
 				<Analytics/>
-				<SpeedInsights/>
 			</body>
 		</html>
 	)