+++ packages/abjad-web/src/app/page.tsx
@@ -10,6 +10,7 @@ const options = [
 	{ value: Abjad.IPA, label: 'IPA' },
 	{ value: Abjad.Arabic, label: 'Arabic ابجد' },
 	{ value: Abjad.HanifiRohingya, label: 'HanifiRohingya 𐴀𐴁𐴅𐴊' },
+	{ value: Abjad.Hebrew, label: 'Hebrew אבגדה' },
 	{ value: Abjad.ImperialAramaic, label: 'Imperial Aramaic 𐡀𐡁𐡂𐡃' },
 	{ value: Abjad.Nabataean, label: 'Nabataean 𐢀𐢂𐢄𐢅' },
 	{ value: Abjad.OldSouthArabian, label: 'Old South Arabian 𐩱𐩨𐩴𐩵' },